Fix db

+++ Nodejs/src/migrations/20241104210836-create-campaign.js
@@ -24,7 +24,7 @@ module.exports = {
         type: Sequelize.TEXT
       },
       goal_amount: {
-        type: Sequelize.DECIMAL(10, 2)
+        type: Sequelize.BIGINT
       },
       start_date: {
         type: Sequelize.DATE
